requirements.txt
yfinance
pandas
numpy
ta
jinja2
pyyaml
